Eliminate abstractTypesMap parameters
We actually only need this in ONE place:
`identifyAbstractTypeForDbType`. Now that it's a static value we can
stop passing it around and importing it all over the codebase!
