+++ mathesar_ui/src/pages/table/TablePage.svelte
@@ -8,7 +8,6 @@
   import LayoutWithHeader from '@mathesar/layouts/LayoutWithHeader.svelte';
   import type { Table } from '@mathesar/models/Table';
   import { makeSimplePageTitle } from '@mathesar/pages/pageTitleUtils';
-  import { abstractTypesMap } from '@mathesar/stores/abstract-types';
   import {
     Meta,
     TabularData,
@@ -40,7 +39,6 @@
   $: tabularData = new TabularData({
     database: table.schema.database,
     table,
-    abstractTypesMap,
     meta,
     shareConsumer,
   });
