--- mathesar_ui/src/stores/abstract-types/type-configs/comboTypes/jsonObjectFactory.ts
@@ -1,13 +1,0 @@
-import { iconUiTypeJsonObject } from '@mathesar/icons';
-
-import type { AbstractTypeConfigurationFactory } from '../../types';
-
-const jsonObjectFactory: AbstractTypeConfigurationFactory = () => ({
-  getIcon: () => iconUiTypeJsonObject,
-  defaultDbType: 'mathesar_types.mathesar_json_object',
-  cellInfo: {
-    type: 'string',
-  },
-});
-
-export default jsonObjectFactory;
