+++ mathesar_ui/src/systems/record-selector/RecordSelectorWindow.svelte
@@ -3,7 +3,6 @@
 
   import { RichText } from '@mathesar/components/rich-text';
   import TableName from '@mathesar/components/TableName.svelte';
-  import { abstractTypesMap } from '@mathesar/stores/abstract-types';
   import { Meta, TabularData } from '@mathesar/stores/table-data';
   import { currentTablesMap } from '@mathesar/stores/tables';
   import Pagination from '@mathesar/utils/Pagination';
@@ -35,7 +34,6 @@
     $tableId && table
       ? new TabularData({
           database: table.schema.database,
-          abstractTypesMap,
           meta: new Meta({ pagination: new Pagination({ size: 10 }) }),
           hasEnhancedPrimaryKeyCell: false,
           table,
