+++ mathesar_ui/src/systems/table-view/table-inspector/record-summary/FkRecordSummaryConfig.svelte
@@ -1,6 +1,5 @@
 <script lang="ts">
   import type { Table } from '@mathesar/models/Table';
-  import { abstractTypesMap } from '@mathesar/stores/abstract-types';
   import { Meta, TabularData } from '@mathesar/stores/table-data';
   import RecordSummaryConfig from '@mathesar/systems/table-view/table-inspector/record-summary/RecordSummaryConfig.svelte';
   import Pagination from '@mathesar/utils/Pagination';
